DISCONTINUED
This Project is discontinued. The architecture and structure is less than optimal and I sadly lack the time and interest to fix this. If somebody would like to create something similar it would be highly recommended starting from scratch because this is mostly hacked together. 

This Repository will be archived as a testament of my former programming skills. Please don't take this as an example of my code!